# Compare equality of label.Label structs directly