@56quarters 2022-03-09
Pushed a commit that uses more realistic label names and values. It changes the benchmarks a bit but they're still about a 14% improvement. I can rebase and update commands to run if desired.